テスト実行時に必要な環境変数のチェックを追加

+++ test/gcs/e2e.sh
@@ -41,6 +41,18 @@ TODAY=`/bin/date +%Y%m%d` # It is used to generate file name to restore
 
 echo "=== $0 started at `/bin/date "+%Y/%m/%d %H:%M:%S"` ==="
 
+# Validate environment variables
+REQUIRED_ENVS=("GCP_ACCESS_KEY_ID" "GCP_SECRET_ACCESS_KEY" "GCP_PROJECT_ID" "TARGET_BUCKET_URL")
+SATISFY=1
+for ((i = 0; i < ${#REQUIRED_ENVS[@]}; i++)) {
+  ENV=$(eval echo "\$${REQUIRED_ENVS[i]}")
+  if [ -z "$ENV" ]; then
+    echo "ERROR: The environment variable ${REQUIRED_ENVS[i]} must be specified in test."
+    SATISFY=0
+  fi
+}
+if [ $SATISFY -ne 1 ]; then exit 1; fi
+
 # Clean up bucket before start mongodb
 docker-compose down -v
